refactor: 로고 이미지 WebP 변환으로 번들 크기 최적화 (#247)
- logo.png → logo.webp 변환
- 이미지 로딩 속도 개선
